Refactor tests to use narrower integer types and update assertions
- Updated various test cases to replace usages of `AsFloat32()` with `AsInt32()` or `ToDouble()` for better type accuracy.
- Adjusted assertions in tests related to numeric literals to reflect the correct expected types (e.g., `sbyte`, `short`, `int`).
- Ensured that struct literal fields are validated against potential narrower integer types after literal narrowing.
- Modified tests for functions and expressions to align with the new type expectations, enhancing overall type safety and correctness.

+++ tests/DatumIngest.Tests/Execution/ExpressionEvaluatorTests.cs
@@ -878,7 +878,7 @@ public class ExpressionEvaluatorTests
         DataValue result = _evaluator.Evaluate(
             new FunctionCallExpression("len", [new ColumnReference("name")]),
             row);
-        Assert.Equal(5f, result.AsFloat32());
+        Assert.Equal(5, result.AsInt32());
     }
 
     [Fact]
