+++ tests/DatumIngest.Tests/Functions/WordCountFunctionTests.cs
@@ -14,49 +14,49 @@ public sealed class WordCountFunctionTests
     public void WordCount_SingleWord_Returns1()
     {
         DataValue result = _function.Execute([DataValue.FromString("hello")]);
-        Assert.Equal(1f, result.AsFloat32());
+        Assert.Equal(1, result.AsInt32());
     }
 
     [Fact]
     public void WordCount_MultipleWords_ReturnsCorrectCount()
     {
         DataValue result = _function.Execute([DataValue.FromString("hello world foo")]);
-        Assert.Equal(3f, result.AsFloat32());
+        Assert.Equal(3, result.AsInt32());
     }
 
     [Fact]
     public void WordCount_EmptyString_Returns0()
     {
         DataValue result = _function.Execute([DataValue.FromString("")]);
-        Assert.Equal(0f, result.AsFloat32());
+        Assert.Equal(0, result.AsInt32());
     }
 
     [Fact]
     public void WordCount_WhitespaceOnly_Returns0()
     {
         DataValue result = _function.Execute([DataValue.FromString("   \t\n  ")]);
-        Assert.Equal(0f, result.AsFloat32());
+        Assert.Equal(0, result.AsInt32());
     }
 
     [Fact]
     public void WordCount_LeadingAndTrailingWhitespace_IgnoresThem()
     {
         DataValue result = _function.Execute([DataValue.FromString("  hello world  ")]);
-        Assert.Equal(2f, result.AsFloat32());
+        Assert.Equal(2, result.AsInt32());
     }
 
     [Fact]
     public void WordCount_MultipleSpacesBetweenWords_CountsCorrectly()
     {
         DataValue result = _function.Execute([DataValue.FromString("a    b\t\tc")]);
-        Assert.Equal(3f, result.AsFloat32());
+        Assert.Equal(3, result.AsInt32());
     }
 
     [Fact]
     public void WordCount_MixedWhitespace_CountsCorrectly()
     {
         DataValue result = _function.Execute([DataValue.FromString("one\ttwo\nthree\r\nfour")]);
-        Assert.Equal(4f, result.AsFloat32());
+        Assert.Equal(4, result.AsInt32());
     }
 
     [Fact]
@@ -84,6 +84,6 @@ public sealed class WordCountFunctionTests
     public void WordCount_ValidateArguments_AcceptsString()
     {
         DataKind result = _function.ValidateArguments([DataKind.String]);
-        Assert.Equal(DataKind.Float32, result);
+        Assert.Equal(DataKind.Int32, result);
     }
 }
